docs: adds a post-card sotry
