+++ storybook/react/src/lib/config/blog.ts
@@ -1,0 +1,30 @@
+import { IPost } from '@/react-blog';
+import { IPageSectionContent } from '@/react-components';
+
+export const articles: IPost[] = [
+  {
+    meta: {
+      slug: 'the-real-reason-you-need-an-icp',
+      categories: ['Product', 'UX'],
+      keywords: [],
+      author: {
+        avatarUrl: '/dave.png',
+        name: 'Dave Benjamin',
+      },
+    },
+    coverImage: { src: '/images/the-real-reason-you-need-an-icp.webp' },
+    title: 'The Real Reason You Need an ICP',
+    subtitle: `You've heard the term, but here's the real reason that not having one is holding you back. `,
+    abstract: [
+      'The term ICP is thrown around a lot, but what does it really mean?',
+      'Here’s the real reason you need an ICP and how to get one.',
+      'You might be surprised at how bad the advice you have been getting is.',
+    ],
+  },
+];
+
+export const pageHeader: IPageSectionContent = {
+  heading: 'Blog',
+  subHeading: 'The latest from our team',
+  body: `We're always learning and growing. Here's what we've been up to lately.`,
+};
